feat: add team crud
