+++ packages/lib/server-only/teams/delete-team.ts
@@ -1,0 +1,9 @@
+'use server';
+
+import { prisma } from '@documenso/prisma';
+
+export const deleteTeam = async (key: { id: string } | { slug: string }) => {
+  return await prisma.team.delete({
+    where: key,
+  });
+};
