+++ packages/lib/server-only/teams/get-team-members.ts
@@ -1,0 +1,16 @@
+'use server';
+
+import { prisma } from '@documenso/prisma';
+
+export const getTeamMembers = async (slug: string) => {
+  return await prisma.teamMember.findMany({
+    where: {
+      team: {
+        slug,
+      },
+    },
+    include: {
+      user: true,
+    },
+  });
+};
