+++ packages/lib/server-only/teams/get-team-roles.ts
@@ -1,0 +1,17 @@
+'use server';
+
+import { prisma } from '@documenso/prisma';
+
+export async function getTeamRoles(userId: number) {
+  const teamRoles = await prisma.teamMember.findMany({
+    where: {
+      userId,
+    },
+    select: {
+      teamId: true,
+      role: true,
+    },
+  });
+
+  return teamRoles;
+}
