+++ packages/lib/server-only/teams/get-team.ts
@@ -1,0 +1,9 @@
+'use server';
+
+import { prisma } from '@documenso/prisma';
+
+export const getTeam = async (key: { id: string } | { slug: string }) => {
+  return await prisma.team.findUniqueOrThrow({
+    where: key,
+  });
+};
