+++ packages/lib/server-only/teams/get-teams.ts
@@ -1,0 +1,20 @@
+'use server';
+
+import { prisma } from '@documenso/prisma';
+
+export const getTeams = async (userId: number) => {
+  return await prisma.team.findMany({
+    where: {
+      members: {
+        some: {
+          userId,
+        },
+      },
+    },
+    include: {
+      _count: {
+        select: { members: true },
+      },
+    },
+  });
+};
