+++ packages/lib/server-only/teams/remove-team-member.ts
@@ -1,0 +1,14 @@
+'use server';
+
+import { prisma } from '@documenso/prisma';
+
+export const removeTeamMember = async (teamId: string, userId: number) => {
+  return await prisma.teamMember.delete({
+    where: {
+      teamId_userId: {
+        teamId,
+        userId,
+      },
+    },
+  });
+};
