--- packages/lib/server-only/teams/team.ts
@@ -1,174 +1,0 @@
-'use server';
-
-import { Role, Team } from '@prisma/client';
-
-import { prisma } from '@documenso/prisma';
-
-export const createTeam = async (param: { userId: number; name: string; slug: string }) => {
-  const { userId, name, slug } = param;
-
-  const team = await prisma.team.create({
-    data: {
-      name,
-      slug,
-    },
-  });
-
-  await addTeamMember(team.id, userId, Role.ADMIN);
-
-  return team;
-};
-
-export const getTeam = async (key: { id: string } | { slug: string }) => {
-  return await prisma.team.findUniqueOrThrow({
-    where: key,
-  });
-};
-
-export const deleteTeam = async (key: { id: string } | { slug: string }) => {
-  return await prisma.team.delete({
-    where: key,
-  });
-};
-
-export const addTeamMember = async (teamId: string, userId: number, role: Role) => {
-  return await prisma.teamMember.upsert({
-    create: {
-      teamId,
-      userId,
-      role,
-    },
-    update: {
-      role,
-    },
-    where: {
-      teamId_userId: {
-        teamId,
-        userId,
-      },
-    },
-  });
-};
-
-export const removeTeamMember = async (teamId: string, userId: number) => {
-  return await prisma.teamMember.delete({
-    where: {
-      teamId_userId: {
-        teamId,
-        userId,
-      },
-    },
-  });
-};
-
-export const getTeams = async (userId: number) => {
-  return await prisma.team.findMany({
-    where: {
-      members: {
-        some: {
-          userId,
-        },
-      },
-    },
-    include: {
-      _count: {
-        select: { members: true },
-      },
-    },
-  });
-};
-
-export async function getTeamRoles(userId: number) {
-  const teamRoles = await prisma.teamMember.findMany({
-    where: {
-      userId,
-    },
-    select: {
-      teamId: true,
-      role: true,
-    },
-  });
-
-  return teamRoles;
-}
-
-// Check if the user is an admin or owner of the team
-export async function isTeamAdmin(userId: number, teamId: string) {
-  const teamMember = await prisma.teamMember.findFirstOrThrow({
-    where: {
-      userId,
-      teamId,
-    },
-  });
-
-  return teamMember.role === Role.ADMIN || teamMember.role === Role.OWNER;
-}
-
-export const getTeamMembers = async (slug: string) => {
-  return await prisma.teamMember.findMany({
-    where: {
-      team: {
-        slug,
-      },
-    },
-    include: {
-      user: true,
-    },
-  });
-};
-
-export const updateTeam = async (slug: string, data: Partial<Team>) => {
-  return await prisma.team.update({
-    where: {
-      slug,
-    },
-    data: data,
-  });
-};
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const isTeamExists = async (condition: any) => {
-  return await prisma.team.count({
-    where: {
-      OR: condition,
-    },
-  });
-};
-
-// Check if the current user has access to the team
-// Should be used in API routes to check if the user has access to the team
-export const throwIfNoTeamAccess = async (userId: number, slug: string) => {
-  if (!userId) {
-    throw new Error('Unauthorized');
-  }
-  const teamMember = await getTeamMember(userId, slug);
-
-  if (!teamMember) {
-    throw new Error('You do not have access to this team');
-  }
-  /// FIXME
-  return {
-    ...teamMember,
-    user: {},
-  };
-};
-
-// Get the current user's team member object
-export const getTeamMember = async (userId: number, slug: string) => {
-  const teamMember = await prisma.teamMember.findFirstOrThrow({
-    where: {
-      userId,
-      team: {
-        slug,
-      },
-      role: {
-        in: ['ADMIN', 'USER', 'OWNER'],
-      },
-    },
-    include: {
-      team: true,
-    },
-  });
-
-  return teamMember;
-};
