+++ packages/lib/server-only/teams/update-team.ts
@@ -1,0 +1,13 @@
+'use server';
+
+import { prisma } from '@documenso/prisma';
+import { Team } from '@documenso/prisma/client';
+
+export const updateTeam = async (slug: string, data: Partial<Team>) => {
+  return await prisma.team.update({
+    where: {
+      slug,
+    },
+    data: data,
+  });
+};
